Move git operations to Git enum, run clean before checkout
Merged GitFix into Git enum — all git operations are now in one place.
checkout(hash:git:) runs prepareRepository first (clean/reset/LFS/submodules),
then checks out the commit. This ensures a clean working tree before
checkout and removes the GitFix class.
