+++ Sources/Files/Files.swift
@@ -62,13 +62,7 @@ public struct Files: Sendable {
 
             Self.logger.debug("Processing commit: \(hash)")
 
-            try await Shell.execute(
-                "git",
-                arguments: ["checkout", hash],
-                workingDirectory: FilePath(repoPath.path(percentEncoded: false))
-            )
-
-            try await GitFix.prepareRepository(git: input.git)
+            try await Git.checkout(hash: hash, git: input.git)
 
             var resultItems: [ResultItem] = []
             for ext in filetypes {
