+++ Sources/Pattern/Pattern.swift
@@ -73,13 +73,7 @@ public struct Pattern: Sendable {
 
             Self.logger.debug("Processing commit: \(hash)")
 
-            try await Shell.execute(
-                "git",
-                arguments: ["checkout", hash],
-                workingDirectory: FilePath(repoPath.path(percentEncoded: false))
-            )
-
-            try await GitFix.prepareRepository(git: input.git)
+            try await Git.checkout(hash: hash, git: input.git)
 
             var resultItems: [ResultItem] = []
             for pattern in patterns {
